pyproject.toml
[tool.poetry]
name = "http_proxy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
mitmproxy = "^8.1.1"
pycryptodome = "^3.15.0"
cryptography = "^37.0.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = {version = "^22.6.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
